exercise 4.21

+++ controllers/blogs.js
@@ -3,14 +3,6 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-// const getTokenFrom = request => {
-//     const authorization = request.get('authorization')
-//     if (authorization && authorization.startsWith('Bearer ')) {
-//         return authorization.replace('Bearer ', '')
-//     }
-//     return null
-// }
-
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
     response.json(blogs)
@@ -19,14 +11,11 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
 
-    // const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
     }
 
-    // const user = await User.findById(body.userId)
     const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
@@ -65,12 +54,22 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    console.log('user', user)
+    const blog = await Blog.findById(request.params.id)
+    console.log('blog', blog)
+
+    if (blog.user.toString() === user._id.toString()) {
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
-    } catch (error) {
-        next(error)
-    }
+    } else return response.status(401).json({ error: "user not authorized" })
+
 })
 
 module.exports = blogsRouter
